+++ Sources/MapboxMaps/SwiftUI/Viewport/Viewport.swift
@@ -89,7 +89,7 @@ public struct Viewport: Equatable {
 
         /// Extra padding that is added for the geometry during bounding box calculation.
         ///
-        /// Note: This different to inset ``Viewport/insetOptions-swift.property``.
+        /// - Note: Geometry padding is different to camera padding ``Viewport/padding-swift.property``.
 #if swift(>=5.8)
     @_documentation(visibility: public)
 #endif
@@ -132,33 +132,13 @@ public struct Viewport: Equatable {
         public var pitch: CGFloat
     }
 
-    /// Represent insets configuration.
-    ///
-    /// Inset configuration is applicable to every kind of viewport configuration except ``Viewport/idle``.
-#if swift(>=5.8)
-    @_documentation(visibility: public)
-#endif
-    public struct InsetOptions: Equatable {
-        /// Insets of viewport.
-#if swift(>=5.8)
-    @_documentation(visibility: public)
-#endif
-        var insets: SwiftUI.EdgeInsets = .init()
-
-        /// Set of edges that which safe area contribution to padding will be ignored.
-#if swift(>=5.8)
-    @_documentation(visibility: public)
-#endif
-        var ignoredSafeAreaEdges: Edge.Set = []
-    }
-
     let storage: Storage
 
-    /// Configures insets of viewport.
+    /// Denotes camera padding of the viewport.
 #if swift(>=5.8)
     @_documentation(visibility: public)
 #endif
-    public var insetOptions = InsetOptions()
+    public var padding = SwiftUI.EdgeInsets()
 
     /// Idle viewport represents the state when user freely drags the map.
     ///
@@ -259,52 +239,40 @@ public struct Viewport: Equatable {
         return Viewport(storage: .followPuck(options))
     }
 
-    /// Creates a new viewport with modified inset options.
+    /// :nodoc:
+    @available(*, deprecated, renamed: "padding")
+    public func inset(by insets: SwiftUI.EdgeInsets, ignoringSafeArea: Edge.Set = []) -> Viewport { self }
+
+    /// :nodoc:
+    @available(*, unavailable, renamed: "padding")
+    public func inset(edges: Edge.Set, length: CGFloat, ignoringSafeArea: Bool = false) -> Viewport { self }
+
+    /// Adds padding to viewport.
     ///
-    /// Insets are ignored for `idle` viewport.
-    ///
-    /// By default, insets are equal to the safe area. This method allows you to set additional insets, or ignore safe area part on the specified edges.
+    /// Safe area insets will be added to that value to form actual padding. If this method called twice, only the last call will take effect.
     ///
     /// - Parameters:
-    ///   - insets: Additional insets, that will be summarized with existing safe area insets.
-    ///   - ignoringSafeArea: A set of edges where safe area's contribution to the resulting inset should be ignored.
+    ///   - padding: Camera
     /// - Returns: A viewport with modified inset options.
 #if swift(>=5.8)
     @_documentation(visibility: public)
 #endif
-    public func inset(by insets: SwiftUI.EdgeInsets, ignoringSafeArea: Edge.Set = []) -> Viewport {
-        var copy = self
-        copy.insetOptions = .init(insets: insets, ignoredSafeAreaEdges: ignoringSafeArea)
-        return copy
+    public func padding(_ padding: SwiftUI.EdgeInsets) -> Viewport {
+        copyAssigned(self, \.padding, padding)
     }
 
-    /// Creates a new MapViewport with modified inset options.
-    ///
-    /// Insets are ignored for `idle` viewport.
-    ///
-    /// By default, insets are equal to the safe area. This method allows you to set additional inset for the specified edges.
-    /// This method can be called multiple times to configure different edges.
+    /// Adds padding to viewport at specific edges.
     ///
     /// - Parameters:
-    ///   - edges: Edges for which to set the additional inset.
-    ///   - length: The length of inset.
-    ///   - ignoringSafeArea: If safe area's contribution should be ignored for the specified edges.
+    ///   - edges: The set of edges to pad. Default is a `all`.
+    ///   - length: Amount of padding in points.
     /// - Returns: A viewport with modified inset options.
 #if swift(>=5.8)
     @_documentation(visibility: public)
 #endif
-    public func inset(edges: Edge.Set, length: CGFloat, ignoringSafeArea: Bool = false) -> Viewport {
+    public func padding(_ edges: Edge.Set = .all, _ length: CGFloat) -> Viewport {
         var copy = self
-
-        for (edge, keyPath) in edgeToInsetMapping where edges.contains(edge) {
-            copy.insetOptions.insets[keyPath: keyPath] = length
-            if ignoringSafeArea {
-                copy.insetOptions.ignoredSafeAreaEdges.insert(edge)
-            } else {
-                copy.insetOptions.ignoredSafeAreaEdges.remove(edge)
-            }
-        }
-
+        copy.padding.updateEdges(edges, length)
         return copy
     }
 
@@ -326,7 +294,7 @@ public struct Viewport: Equatable {
 
     /// Returns the camera options if viewport is configured with camera options.
     ///
-    /// - Note: The ``CameraOptions-swift.struct/padding`` is ignored, it is replaced with ``Viewport/insetOptions-swift.property``, see ``inset(by:ignoringSafeArea:)``.
+    /// - Note: The ``CameraOptions-swift.struct/padding`` is ignored, it is replaced with ``Viewport/padding``, see ``Viewport/padding(_:)``.
 #if swift(>=5.8)
     @_documentation(visibility: public)
 #endif
@@ -369,38 +337,22 @@ public struct Viewport: Equatable {
 @available(iOS 13.0, *)
 extension Viewport {
     func makeState(with mapView: MapView, layoutDirection: LayoutDirection) -> ViewportState? {
-        // TODO: mapView's safeAreaInsets don't reflect the real safe area added by SwiftUI views (such as toolbars).
-        let insets = padding(with: layoutDirection, safeAreaInsets: mapView.safeAreaInsets)
+        let padding = UIEdgeInsets(insets: padding, layoutDirection: layoutDirection)
         switch storage {
         case .idle:
             return nil
         case .camera(var cameraOptions):
-            cameraOptions.padding = insets
-            return CameraViewportState(cameraOptions: cameraOptions)
+            cameraOptions.padding = padding
+            return mapView.viewport.makeCameraViewportState(camera: cameraOptions)
         case .overview(let options):
-            let options = options.resolve(layoutDirection: layoutDirection, padding: insets)
+            let options = options.resolve(layoutDirection: layoutDirection, padding: padding)
             return mapView.viewport.makeOverviewViewportState(options: options)
         case .styleDefault:
-            return DefaultStyleViewportState(
-                mapboxMap: mapView.mapboxMap,
-                styleManager: mapView.mapboxMap,
-                padding: insets)
+            return mapView.viewport.makeDefaultStyleViewportState(padding: padding)
         case .followPuck(let options):
-            let options = options.resolve(padding: insets)
+            let options = options.resolve(padding: padding)
             return mapView.viewport.makeFollowPuckViewportState(options: options)
         }
-    }
-
-    func padding(with layoutDirection: LayoutDirection, safeAreaInsets: UIEdgeInsets) -> UIEdgeInsets {
-        var result = SwiftUI.EdgeInsets(uiInsets: safeAreaInsets, layoutDirection: layoutDirection)
-
-        for (edge, keyPath) in edgeToInsetMapping where insetOptions.ignoredSafeAreaEdges.contains(edge) {
-            result[keyPath: keyPath] = 0
-        }
-
-        result += insetOptions.insets
-
-        return UIEdgeInsets(insets: result, layoutDirection: layoutDirection)
     }
 }